change type item quantity to int 64

+++ be/dto/item_dto.go
@@ -9,7 +9,7 @@ import (
 type ItemCreationRequest struct {
 	Name        string  `json:"name" binding:"required"`
 	Description string  `json:"description"`
-	Quantity    int     `json:"quantity"`
+	Quantity    int64   `json:"quantity"`
 	Price       float64 `json:"price"`
 	Picture     string  `json:"picture"`
 }
@@ -18,7 +18,7 @@ type ItemCreationResponse struct {
 	ID          uuid.UUID `json:"id"`
 	Name        string    `json:"name" binding:"required"`
 	Description string    `json:"description"`
-	Quantity    int       `json:"quantity"`
+	Quantity    int64     `json:"quantity"`
 	Price       float64   `json:"price"`
 	Picture     string    `json:"picture"`
 	CreatedAt   time.Time `json:"created_at"`
@@ -27,7 +27,7 @@ type ItemCreationResponse struct {
 type ItemUpdatingRequest struct {
 	Name        string  `json:"name" binding:"required"`
 	Description string  `json:"description"`
-	Quantity    int     `json:"quantity"`
+	Quantity    int64   `json:"quantity"`
 	Price       float64 `json:"price"`
 	Picture     string  `json:"picture"`
 }
@@ -35,7 +35,7 @@ type ItemUpdatingRequest struct {
 type ItemUpdatingResponse struct {
 	Name        string    `json:"name"`
 	Description string    `json:"description"`
-	Quantity    int       `json:"quantity"`
+	Quantity    int64     `json:"quantity"`
 	Price       float64   `json:"price"`
 	Picture     string    `json:"picture"`
 	UpdatedAt   time.Time `json:"updated_at"`
